projet/pyproject.toml
[project]
name = "eth-correlation-graph"
version = "0.1.0"
description = "Ethereum Address Correlation calculation using NetworkX and Dune Analytics"
authors = [
    { name = "Your Name", email = "you@example.com" }
]
requires-python = ">=3.10"
dependencies = [
    "networkx>=3.1",
    "pandas>=2.0",
    "dune-client>=1.0",
    "python-dotenv>=1.0",
    "matplotlib>=3.7",
    "scipy>=1.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "black>=23.0",
]
